chore(web-client): move tests to separate directory
