+++ web-client/tsconfig.test.json
@@ -7,6 +7,7 @@
     "esModuleInterop": true
   },
   "include": [
-    "src"
+    "src",
+    "test"
   ]
 }
